Move optical sensor code to drivers folder
