+++ keyboards/handwired/dactyl_manuform/5x6_right_trackball/5x6_right_trackball.h
@@ -19,7 +19,7 @@
 #include "dactyl_manuform.h"
 #include "quantum.h"
 #include "spi_master.h"
-#include "pmw3360.h"
+#include "drivers/sensors/pmw3360.h"
 #include "pointing_device.h"
