+++ keyboards/handwired/dactyl_manuform/5x6_right_trackball/rules.mk
@@ -27,5 +27,5 @@ MOUSE_SHARED_EP = no
 SPLIT_KEYBOARD = yes
 SPLIT_TRANSPORT = custom
 
-SRC += pmw3360.c
+SRC += drivers/sensors/pmw3360.c
 QUANTUM_LIB_SRC += pointer_transport.c serial.c i2c_master.c i2c_slave.c spi_master.c
