+++ keyboards/ploopyco/trackball/trackball.h
@@ -20,7 +20,7 @@
 
 #include "quantum.h"
 #include "spi_master.h"
-#include "pmw3360.h"
+#include "drivers/sensors/pmw3360.h"
 #include "analog.h"
 #include "opt_encoder.h"
 #include "pointing_device.h"
