+++ keyboards/ploopyco/trackball_mini/trackball_mini.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include "quantum.h"
-#include "adns5050.h"
+#include "drivers/sensors/adns5050.h"
 #include "analog.h"
 #include "opt_encoder.h"
 #include "pointing_device.h"
