+++ keyboards/ploopyco/trackball_nano/trackball_nano.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include "quantum.h"
-#include "adns5050.h"
+#include "drivers/sensors/adns5050.h"
 #include "analog.h"
 #include "opt_encoder.h"
 #include "pointing_device.h"
